requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.8.3=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.3.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=3.6.2=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.2.1=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.2.0=pypi_0
babel=2.11.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bleach=5.0.1=pypi_0
blis=0.7.7=pypi_0
brotlipy=0.7.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.12.7=ha878542_0
catalogue=2.0.8=pypi_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=pypi_0
cfgv=3.3.1=pypi_0
charset-normalizer=2.1.1=pypi_0
clearml=1.9.0=pypi_0
click=8.0.4=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.2=pypi_0
contourpy=1.0.7=pypi_0
coverage=7.0.5=pypi_0
cryptography=38.0.4=pypi_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
cymem=2.0.6=pypi_0
cython-blis=0.7.7=py310h9102076_0
dataclasses=0.8=pyhc8e2a94_3
debugpy=1.6.5=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
distlib=0.3.6=pypi_0
en-core-web-md=3.2.0=pypi_0
en-core-web-sm=3.2.0=pypi_0
entrypoints=0.4=pypi_0
executing=1.2.0=pypi_0
fastjsonschema=2.16.2=pypi_0
fasttext=0.9.2=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=pypi_0
fsspec=2022.11.0=pypi_0
fst-pso=1.8.1=pypi_0
furl=2.1.3=pypi_0
fuzzytm=2.0.5=pypi_0
gensim=4.3.0=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
huggingface-hub=0.11.1=pypi_0
hydra-core=1.3.1=pypi_0
identify=2.5.17=pypi_0
idna=3.4=pypi_0
imageio=2.24.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.20.1=pypi_0
ipython=8.8.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.0.4=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pyhd8ed1ab_1
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
json5=0.9.11=pypi_0
jsonpointer=2.3=pypi_0
jsonschema=4.17.3=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.4.9=pypi_0
jupyter-console=6.4.4=pypi_0
jupyter-core=5.1.3=pypi_0
jupyter-events=0.6.3=pypi_0
jupyter-server=2.1.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab=3.5.2=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.18.0=pypi_0
jupyterlab-widgets=3.0.5=pypi_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
langcodes=3.3.0=pyhd8ed1ab_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=hecacb30_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lightning-utilities=0.5.0=pypi_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=pypi_0
matplotlib=3.6.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
miniful=0.0.6=pypi_0
mistune=2.0.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-fft=1.3.1=pypi_0
mkl-random=1.2.2=pypi_0
mkl-service=2.4.0=pypi_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
multidict=6.0.4=pypi_0
murmurhash=1.0.7=pypi_0
nbclassic=0.4.8=pypi_0
nbclient=0.7.2=pypi_0
nbconvert=7.2.7=pypi_0
nbformat=5.7.3=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.6=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.0=pypi_0
nltk=3.8.1=pypi_0
nodeenv=1.7.0=pypi_0
notebook=6.5.2=pypi_0
notebook-shim=0.2.2=pypi_0
numpy=1.24.1=pypi_0
numpy-base=1.23.5=py310h8e6c178_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.7.0.68=pypi_0
opencv-python-headless=4.7.0.68=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
orderedmultidict=1.0.1=pypi_0
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.2=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathlib2=2.3.7.post1=pypi_0
pathy=0.10.1=pyhd8ed1ab_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=pypi_0
platformdirs=2.6.2=pypi_0
pre-commit=3.0.2=pypi_0
preshed=3.0.6=pypi_0
prometheus-client=0.15.0=pypi_0
prompt-toolkit=3.0.36=pypi_0
protobuf=3.20.1=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pybind11=2.10.3=pypi_0
pycocotools=2.0.6=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=pypi_0
pydicom=2.3.1=pypi_0
pyfume=0.2.25=pypi_0
pygments=2.14.0=pypi_0
pyjwt=2.4.0=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=pypi_0
pysocks=1.7.1=pypi_0
python=3.10.8=h7a1cb2a_1
python-dateutil=2.8.2=pypi_0
python-dotenv=0.21.1=pypi_0
python-json-logger=2.0.4=pypi_0
python_abi=3.10=2_cp310
pytorch=1.12.1=py3.10_cuda10.2_cudnn7.6.5_0
pytorch-lightning=1.8.6=pypi_0
pytorch-mutex=1.0=cuda
pytz=2022.7.1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.0.0=pypi_0
qtconsole=5.4.0=pypi_0
qtpy=2.3.0=pypi_0
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
regex=2022.10.31=pypi_0
requests=2.28.1=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.2.0=pypi_0
scipy=1.10.0=pypi_0
send2trash=1.8.0=pypi_0
setuptools=65.6.3=pypi_0
shellingham=1.5.0.post1=pyhd8ed1ab_0
simpful=2.9.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
smart-open=6.3.0=pypi_0
smart_open=5.2.1=pyhd8ed1ab_0
sniffio=1.3.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
spacy=3.2.0=pypi_0
spacy-legacy=3.0.11=pyhd8ed1ab_0
spacy-loggers=1.0.4=pyhd8ed1ab_0
sqlite=3.40.1=h5082296_0
srsly=2.4.3=pypi_0
stack-data=0.6.2=pypi_0
tensorboardx=2.5.1=pypi_0
terminado=0.17.1=pypi_0
thinc=8.0.15=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.10.10=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.2=pypi_0
tomli=2.0.1=pypi_0
torch=1.12.1=pypi_0
torchaudio=0.12.1=pypi_0
torchmetrics=0.11.0=pypi_0
torchtext=0.13.0=pypi_0
torchvision=0.13.1=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.1=pyhd8ed1ab_0
traitlets=5.8.1=pypi_0
transformers=4.25.1=pypi_0
typer=0.4.2=pyhd8ed1ab_0
typing-extensions=4.4.0=pypi_0
typing_extensions=4.4.0=py310h06a4308_0
tzdata=2022g=h04d1e81_0
uri-template=1.2.0=pypi_0
urllib3=1.26.13=pypi_0
virtualenv=20.17.1=pypi_0
wasabi=0.10.1=pyhd8ed1ab_1
wcwidth=0.2.6=pypi_0
webcolors=1.12=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.4.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.5=pypi_0
xz=5.2.8=h5eee18b_0
yarl=1.8.2=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
